Convert oracle_resync and ethereum_event_watcher

+++ orchestrator/gorc/src/commands/orchestrator/start.rs
@@ -97,7 +97,7 @@ impl Runnable for StartCommand {
                 gas_price,
                 &config.metrics.listen_addr,
                 config.ethereum.gas_price_multiplier,
-                config.ethereum.blocks_to_search as u128,
+                config.ethereum.blocks_to_search,
                 config.cosmos.gas_adjustment,
                 self.orchestrator_only,
                 config.cosmos.msg_batch_size,
